requirements.txt
earthengine-api>=1.5.0
geopandas>=0.10.0
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0
shapely>=1.8.0
qdrant-client>=1.0.0
scikit-learn>=1.0.0
pillow>=9.0.0
h3>=4.0.0
tqdm>=4.60.0
folium>=0.12.0
rasterio>=1.2.0
torch>=1.10.0
torchvision>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0